+++ apps/matlab/Matlab_gt_deformation_field_reg_2D_series.cpp
@@ -229,7 +229,7 @@ void mexFunction(int nlhs,mxArray *plhs[],int nrhs,const mxArray *prhs[])
         hoNDImageContainer2D<Image2DType> imageSeries;
         imageSeries.create(target.begin(), dim);
 
-        Gadgetron::hoImageRegContainer2DRegistration<Image2DType, Image2DType, float> regContainer;
+        Gadgetron::hoImageRegContainer2DRegistration<Image2DType, Image2DType, double> regContainer;
 
         regContainer.setDefaultParameters(level, false);
 
@@ -303,7 +303,7 @@ void mexFunction(int nlhs,mxArray *plhs[],int nrhs,const mxArray *prhs[])
         // output parameter
         // ---------------------------------------------------------------
 
-        hoNDArray<float> deformField;
+        hoNDArray<double> deformField;
 
         mxArray* M0 = NULL;
         regContainer.deformation_field_[0].to_NDArray(0, deformField);
